storage: clean unrelated to bsb files
